+++ Assets/PlayMaker uGui/Proxies/PlayMakerUGuiCanvasRaycastFilterEventsProxy.cs
@@ -1,0 +1,24 @@
+﻿using UnityEngine;
+using UnityEngine.UI;
+using UnityEngine.EventSystems;
+using System.Collections;
+
+using HutongGames.PlayMaker;
+using HutongGames.PlayMaker.Actions;
+using HutongGames.PlayMaker.Ecosystem.Utils;
+
+public class PlayMakerUGuiCanvasRaycastFilterEventsProxy : MonoBehaviour, ICanvasRaycastFilter
+{
+
+	public bool RayCastingEnabled = true;
+
+	#region ICanvasRaycastFilter implementation
+	public bool IsRaycastLocationValid (Vector2 sp, Camera eventCamera)
+	{
+		return RayCastingEnabled;
+	}
+	#endregion
+
+
+	
+}